Webservice: only create WATCH_FN if it does not exist yet
If we always touch() the WATCH_FN, deploying a new version in parallel
to a running version (attemping to do an upgrade without much downtime)
will always cause the already running version to trigger it's updater().
But especially during such an upgrade process, that puts unnecessary
load on the machine.